implemented classes
This is the most updated version of main.py (the final commit). GitHub has been giving me issues again with committing, and I'd rather not risk taking much more time trying to fix it because I want this submission to be on time. I will keep the other main.py on the repository, though, so that you can see my commit history.